moved invspinctrl to widgets
